+++ Changelog.md
@@ -1,3 +1,7 @@
+## 1.0.0
+- Removing status transition & only adding a comment to the issue
+- The comment consists of a quote of the commits deployed (based upon the JIRA key in the commit message)
+
 ## 0.3.0
 - Filter transited issues by commit messages